requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics==8.1.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
